Remove to-vfile and vfile-reporter dependencies
unified uses an outdated vfile@5.0.0 that's not compatible with to-vfile@8.0.0.
Also, it's not necessary - we can read/write with Node fs. vfile-reporter _does_
need a vfile as input, but we're not using it to check anything, so it always
prints "no issues found". From the docs: "vfile [is a] utility to create a report".
If we add prose checking or such, we can readd it later.